backend/pyproject.toml
[project]
name = "attenborofy"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask>=2.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=0.19.0",
    "elevenlabs>=2.40.0",
    "openai>=1.0.0",
    "gunicorn>=21.2.0",
    "azure-storage-blob>=12.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
